OpenZFS: use ptraddr_t when we want addresses
Whenever we want to manipulate (or store) the address of the pointer,
use ptraddr_t, not uintptr_t.  In CHERI C, the latter is in integer
pointer address while uintptr_t may carry metadata as required to
permit a void * to be stored and turned back into a usable void *
later.